+++ Assets/Scripts/GameManager.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class GameManager : MonoBehaviour
 {
-    public int life; //vie du bateau
+    public float life; //vie du bateau
     public List<GameObject> problems;
     public float BaseTimeToWait;
